xPollApi/requirements.txt
coverage==6.3.2                                                                                            
Django==4.0.2                                                                                              
django-heroku==0.3.1                                                                                       
djangorestframework==3.13.1                                                                               
gunicorn==20.1.0                                                                                          
Pillow==9.0.1                                                                                             
psycopg2==2.9.3                                                                                           
psycopg2-binary==2.9.3                                                                                    
whitenoise==5.3.0           
